+++ cli/soit_cli/__main__.py
@@ -1,0 +1,5 @@
+import sys
+
+from soit_cli.main import main
+
+sys.exit(main())